move some functions into util so it can be tested, and move utilities for testing into common.py
