+++ conftest.py
@@ -3,7 +3,7 @@ import pytest
 # Register tests/test_utils.py module as a pytest plugin so that tests can use fixtures from that
 # module without importing them (pytest recommends against importing them, and linters will flag the
 # import as unused and the fixture usage in test method args as redefining the var; https://github.com/astral-sh/ruff/issues/4046)
-pytest_plugins = "tests.test_utils"
+pytest_plugins = "tests.common"
 
 
 def pytest_addoption(parser):
