+++ tests/test_api_client.py
@@ -5,13 +5,15 @@ import re
 import unittest
 from pathlib import Path
 from unittest.mock import MagicMock, Mock, PropertyMock, call
+from collections import namedtuple
+from datetime import date, timedelta, datetime
 
 import pendulum
 import pytest
 import requests
 from sqlalchemy.orm import Session
 
-from tests.test_utils import (
+from tests.common import (
     all_crawls,
     all_videos,
 )
@@ -822,3 +824,4 @@ def test_TikTokVideoRequest_as_json(basic_video_query_config):
         "search_id": None,
         "start_date": "20240601",
     }
+
